+++ tests/test_graphics_images.py
@@ -3,18 +3,23 @@
 """
 Tests for RLG Image shapes.
 """
-from reportlab.lib.testutils import setOutDir,makeSuiteForClasses, outputfile, printLocation
+from reportlab.lib.testutils import setOutDir,makeSuiteForClasses, outputfile, printLocation, rlSkipUnless
 setOutDir(__name__)
 import os
 import unittest
-from reportlab.graphics.shapes import Image, Drawing, Rect
+from reportlab.graphics.shapes import Image, Drawing, Rect, Group, Line
 from reportlab.graphics import renderPDF
 from reportlab.lib.pagesizes import A4
-from reportlab.lib.colors import toColor
+from reportlab.lib.colors import toColor, CMYKColor, rgb2cmyk
 
 IMAGENAME = 'pythonpowered.gif'
 GSIMAGE = 'pythonpowered-gs.gif'
 GAIMAGE = 'gray-alpha.png'
+
+try:
+    import rlPyCairo
+except ImportError:
+    rlPyCairo = None
 
 class ImageTestCase(unittest.TestCase):
     "Test RLG Image shape."
@@ -32,11 +37,6 @@ class ImageTestCase(unittest.TestCase):
         outPath = outputfile("test_graphics_images.pdf")
         renderPDF.drawToFile(d, outPath) #, '')
         assert os.path.exists(outPath)
-
-        try:
-            import rlPyCairo
-        except ImportError:
-            rlPyCairo = None
 
         try:
             import _rl_renderPM
@@ -119,6 +119,72 @@ class ImageTestCase(unittest.TestCase):
         addImage(10,10,110,44)
         d.translate(72,6*72)
         self.IMAGES.append(d)
+
+    @rlSkipUnless(rlPyCairo,'cannot import rlPyCairo')
+    def test7(self):
+        rw = 10
+        rg = 10
+        xgap = 5
+        h = w = 3*rw+4*rg
+        d = Drawing(2*h + xgap,h)
+        C = toColor('red'),toColor('green'),toColor('blue')
+        g = Group()
+
+        def colconv(j,c):
+            if j==0: return c
+            if j==1: return CMYKColor(*rgb2cmyk(c.red,c.green,c.blue),alpha=c.alpha)
+
+        from reportlab.graphics.widgets.grids import ShadedRect
+        for i in range(3):
+            c = C[i]
+            sc = c.clone(alpha=0.1)
+            ec = c.clone(alpha=1.0)
+            for j in range(2):
+                x = j*(w+xgap)
+                d.add(ShadedRect(x=x,y=rg+(rg+rw)*i,width=w,height=rw,orientation='vertical',numShades=w//10,
+                        fillColorStart=colconv(j,sc),fillColorEnd=colconv(j,ec),
+                        strokeColor=None,strokeWidth=0))
+
+                g.add(ShadedRect(x=x+rg+(rg+rw)*i,y=0,width=rw,height=h,orientation='horizontal',numShades=w//10,
+                        fillColorStart=colconv(j,sc),fillColorEnd=colconv(j,ec),
+                        strokeColor=None,strokeWidth=0))
+        d.add(g)
+
+        d._renderPM_backendFmt = 'ARGB32'
+        d._renderPM_bg = toColor('white').clone(alpha=0)    #transparent white
+        d.save(formats=['png'],
+            fnRoot=outputfile(os.path.join("charts-out","test_graphics_images_alpha_test.png")), outDir=None)
+
+    @rlSkipUnless(rlPyCairo,'cannot import rlPyCairo')
+    def test8(self):
+        rw = 10
+        rg = 10
+        sw = 2
+        xgap = 5
+        h = w = 3*rw+4*rg
+        d = Drawing(2*h + xgap,h)
+        C = toColor('red'),toColor('green')
+        g = Group()
+
+        def colconv(j,c):
+            if j==0: return c
+            if j==1: return CMYKColor(*rgb2cmyk(c.red,c.green,c.blue),alpha=c.alpha)
+
+        for i in range(3):
+            ch = C[0].clone(alpha=0.1 + 0.45*i)
+            cv = C[1].clone(alpha=0.1 + 0.45*i)
+            for j in range(2):
+                x = j*(w+xgap)
+                y = rg+(rg+rw)*i + rw/2
+                d.add(Line(x,y,x+w,y, strokeColor=ch, strokeWidth=sw))
+                x += rg+(rg+rw)*i+ rw/2
+                g.add(Line(x,0,x, h, strokeColor=cv, strokeWidth=sw))
+        d.add(g)
+
+        d._renderPM_backendFmt = 'ARGB32'
+        d._renderPM_bg = toColor('white').clone(alpha=0)    #transparent white
+        d.save(formats=['png'],
+            fnRoot=outputfile(os.path.join("charts-out","test_graphics_images_alpha_test8.png")), outDir=None)
 
 def makeSuite():
     return makeSuiteForClasses(ImageTestCase)